Depth Estimation/requirements.txt
transformers
Pillow
gradio
torch
torchvision